requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
umap-learn
nltk
plotly
streamlit
python-dotenv
